+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Changed password reuse error response code from 400 to 409
 - Changed expired method verification response code from 400 to 410
 - /auth/login returns 400 for client_id missing, instead of 302
+- Added `uuid` property to `/user/me` response
 - Validation attempt on expired reset now issues a new reset
 ### Removed
 - Removed support for phone password recovery methods